requirements.txt
python-frontmatter
PyYAML
pyvips[binary]
matplotlib
toml
